pyproject.toml
[project]
name = "dashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "arize>=8.4.0",
    "fastapi>=0.133.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.41.0",
]
